ngfw-15866 reverted captive portal changes and renamed the file name
