+++ untangle-vue-ui/source/src/components/settings/services/IntrusionPrevesion.vue
@@ -51,7 +51,7 @@
   import { mapGetters } from 'vuex'
   import { IntrusionPrevention, NoLicense, UAppStatusRemove, UAppInstall, intrusionPreventionAllTabs } from 'vuntangle'
   import { VDivider } from 'vuetify/lib'
-  import { ngfwCapabilities } from './IPCapabilities'
+  import { ngfwCapabilities } from './IpCapabilities'
   import serviceMixin from './serviceMixin'
 
   export default {
